Update to 391e3074b3447b3a0d5f9153eb82cb87c89dfede (1.98a) about bug fixes.

+++ getput.c
@@ -2019,7 +2019,10 @@ static void DispDownloadFinishMsg(TRANSPACKET *Pkt, int iRetCode)
 //			else if((Pkt->hWndTrans != NULL) && (TimeStart != 0))
 //				SetTaskMsg(MSGJPN102, TimeStart, Pkt->ExistSize/TimeStart);
 			else if((Pkt->hWndTrans != NULL) && (TimeStart[Pkt->ThreadCount] != 0))
-				SetTaskMsg(MSGJPN102, TimeStart[Pkt->ThreadCount], Pkt->ExistSize/TimeStart[Pkt->ThreadCount]);
+				// "0 B/S"と表示されるバグを修正
+				// 原因は%dにあたる部分に64ビット値が渡されているため
+//				SetTaskMsg(MSGJPN102, TimeStart[Pkt->ThreadCount], Pkt->ExistSize/TimeStart[Pkt->ThreadCount]);
+				SetTaskMsg(MSGJPN102, (LONG)TimeStart[Pkt->ThreadCount], (LONG)(Pkt->ExistSize/TimeStart[Pkt->ThreadCount]));
 			else
 				SetTaskMsg(MSGJPN103, Pkt->ExistSize);
 		}
